+++ snapcraft/internal/errors.py
@@ -14,17 +14,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import contextlib
-
-from snapcraft import formatting_utils
-
-# dict of jsonschema validator -> cause pairs. Wish jsonschema just gave us
-# better messages.
-_VALIDATION_ERROR_CAUSES = {
-    'maxLength': 'maximum length is {validator_value}',
-    'minLength': 'minimum length is {validator_value}',
-}
-
 
 class SnapcraftError(Exception):
     """Base class for all snapcraft exceptions.
@@ -71,17 +60,6 @@ class PrimeFileConflictError(SnapcraftError):
     )
 
 
-class DuplicateAliasError(SnapcraftError):
-
-    fmt = 'Multiple parts have the same alias defined: {aliases!r}'
-
-    def __str__(self):
-        if isinstance(self.aliases, (list, set)):
-            self.aliases = ','.join(self.aliases)
-
-        return super().__str__()
-
-
 class InvalidAppCommandError(SnapcraftError):
 
     fmt = (
@@ -122,14 +100,6 @@ class PartNotInCacheError(SnapcraftError):
     )
 
 
-class SnapcraftLogicError(SnapcraftError):
-
-    fmt = 'Issue detected while analyzing snapcraft.yaml: {message}'
-
-    def __init__(self, message):
-        super().__init__(message=message)
-
-
 class PluginError(SnapcraftError):
 
     fmt = 'Issue while loading part: {message}'
@@ -142,13 +112,6 @@ class PluginNotDefinedError(SnapcraftError):
 
     fmt = ("Issues while validating snapcraft.yaml: the 'plugin' keyword is "
            "missing for the {part_name} part.")
-
-
-class SnapcraftYamlFileError(SnapcraftError):
-
-    fmt = ('Could not find {snapcraft_yaml}. Are you sure you are in the '
-           'right directory?\n'
-           'To start a new project, use `snapcraft init`')
 
 
 class SnapcraftPartConflictError(SnapcraftError):
@@ -228,83 +191,3 @@ class RequiredPathDoesNotExist(SnapcraftError):
 class SnapcraftPathEntryError(SnapcraftError):
 
     fmt = 'The path {value!r} set for {key!r} in {app!r} does not exist.'
-
-
-class SnapcraftSchemaError(SnapcraftError):
-
-    fmt = 'Issues while validating {snapcraft_yaml}: {message}'
-
-    @classmethod
-    def from_validation_error(cls, error):
-        """Take a jsonschema.ValidationError and create a SnapcraftSchemaError.
-
-        The validation errors coming from jsonschema are a nightmare. This
-        class tries to make them a bit more understandable.
-        """
-
-        messages = []
-
-        # error.validator_value may contain a custom validation error message.
-        # If so, use it instead of the garbage message jsonschema gives us.
-        with contextlib.suppress(TypeError, KeyError):
-            messages.append(
-                error.validator_value['validation-failure'].format(error))
-
-        if not messages:
-            messages.append(error.message)
-
-        path = []
-        while error.absolute_path:
-            element = error.absolute_path.popleft()
-            # assume numbers are indices and use 'xxx[123]' notation.
-            if isinstance(element, int):
-                path[-1] = '{}[{}]'.format(path[-1], element)
-            else:
-                path.append(str(element))
-        if path:
-            messages.insert(0, "The '{}' property does not match the "
-                               "required schema:".format('/'.join(path)))
-        cause = error.cause or _determine_cause(error)
-        if cause:
-            messages.append('({})'.format(cause))
-
-        return cls(' '.join(messages))
-
-    def __init__(self, message, snapcraft_yaml='snapcraft.yaml'):
-        super().__init__(message=message, snapcraft_yaml=snapcraft_yaml)
-
-
-def _determine_cause(error):
-    """Attempt to determine a cause from validation error.
-
-    :return: A string representing the cause of the error (it may be empty if
-             no cause can be determined).
-    :rtype: str
-    """
-
-    message = _VALIDATION_ERROR_CAUSES.get(error.validator, '').format(
-        validator_value=error.validator_value)
-
-    if not message and error.validator == 'anyOf':
-        message = _interpret_anyOf(error)
-
-    return message
-
-
-def _interpret_anyOf(error):
-    """Interpret a validation error caused by the anyOf validator.
-
-    Returns:
-        A string containing a (hopefully) helpful validation error message. It
-        may be empty.
-    """
-
-    usages = []
-    try:
-        for validator in error.validator_value:
-            usages.append(validator['usage'])
-    except (TypeError, KeyError):
-        return ''
-
-    return 'must be one of {}'.format(formatting_utils.humanize_list(
-        usages, 'or'))
