+++ snapcraft/tests/project_loader/test_parts.py
@@ -43,11 +43,17 @@ class TestParts(tests.TestCase):
         self.mock_get_yaml.return_value = os.path.join(
             'snap', 'snapcraft.yaml')
         self.addCleanup(patcher.stop)
+
+        patcher = unittest.mock.patch(
+            'snapcraft.internal.project_loader._parts_config.PartsConfig'
+            '.load_plugin')
+        self.mock_plugin_loader = patcher.start()
+        self.addCleanup(patcher.stop)
+
         self.part_schema = project_loader.Validator().part_schema
         self.deb_arch = snapcraft.ProjectOptions().deb_arch
 
-    @unittest.mock.patch('snapcraft.internal.parts.PartsConfig.load_plugin')
-    def test_get_parts_none(self, mock_loadPlugin):
+    def test_get_parts_none(self):
         self.make_snapcraft_yaml("""name: test
 version: "1"
 summary: test
@@ -62,8 +68,7 @@ parts:
         config = project_loader.load_config(None)
         self.assertThat(config.parts.get_part('not-a-part'), Equals(None))
 
-    @unittest.mock.patch('snapcraft.internal.parts.PartsConfig.load_plugin')
-    def test_slash_warning(self, mock_loadPlugin):
+    def test_slash_warning(self):
         fake_logger = fixtures.FakeLogger(level=logging.WARN)
         self.useFixture(fake_logger)
 
@@ -84,8 +89,7 @@ parts:
             'DEPRECATED: Found a "/" in the name of the {!r} part'.format(
                 'part/1')))
 
-    @unittest.mock.patch('snapcraft.internal.parts.PartsConfig.load_plugin')
-    def test_snap_deprecation(self, mock_loadPlugin):
+    def test_snap_deprecation(self):
         """Test that using the 'snap' keyword results in a warning."""
 
         fake_logger = fixtures.FakeLogger(level=logging.WARN)
